backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
httpx>=0.26.0
sqlalchemy>=2.0.0
redis>=5.0.0
apscheduler>=3.10.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.1
email-validator>=2.1.1
